LayoutTests/storage/indexeddb/modern/idbobjectstore-count-1-expected.txt

This test exercises various uses of IDBObjectStore.count()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Count is: 1
ALERT: Count is: 0
ALERT: Count is: 0
ALERT: Count is: 0
ALERT: Count is: 0
ALERT: Count is: 2
ALERT: Count is: 0
ALERT: Count is: 0
ALERT: Count is: 0
ALERT: Count is: 0
ALERT: Count is: 3
ALERT: Count is: 1
ALERT: Count is: 0
ALERT: Count is: 1
ALERT: Count is: 0
ALERT: Count is: 4
ALERT: Count is: 2
ALERT: Count is: 1
ALERT: Count is: 2
ALERT: Count is: 0
ALERT: Count is: 5
ALERT: Count is: 3
ALERT: Count is: 2
ALERT: Count is: 3
ALERT: Count is: 0
ALERT: Count is: 6
ALERT: Count is: 4
ALERT: Count is: 3
ALERT: Count is: 3
ALERT: Count is: 0
ALERT: Count is: 7
ALERT: Count is: 4
ALERT: Count is: 3
ALERT: Count is: 3
ALERT: Count is: 1
ALERT: Count is: 8
ALERT: Count is: 4
ALERT: Count is: 3
ALERT: Count is: 3
ALERT: Count is: 1
ALERT: Count is: 9
ALERT: Count is: 4
ALERT: Count is: 3
ALERT: Count is: 3
ALERT: Count is: 1
ALERT: Count is: 10
ALERT: Count is: 4
ALERT: Count is: 3
ALERT: Count is: 3
ALERT: Count is: 1
ALERT: Initial upgrade versionchange transaction complete
PASS successfullyParsed is true

TEST COMPLETE
